Alzheimer's disease assays

Returns a tibble containing curated metadata for assay implementations used to measure fluid or imaging biomarkers relevant to Alzheimer's disease research. Each row describes a specific combination of vendor, platform, and technology used to quantify a given analyte.

Assays represent product-specific implementations (e.g., Elecsys, Simoa HD-X, INNOTEST) that produce numeric measurements for downstream biomarker classification. Quantitative performance parameters such as lower and upper limits of quantification (LLOQ, ULOQ) are included where available.

Usage

assays()

Value

A tibble with the following columns:

assay_id

Unique identifier for the assay.

analyte_id

Reference to the measured analyte (see analytes()).

modality

Either "fluid" or "imaging", denoting the type of measurement.

matrix

Biological specimen used, such as "CSF", "plasma", or "brain".

lloq

Lower limit of quantification, in appropriate units.

uloq

Upper limit of quantification.

unit

Measurement unit (e.g., "pg/mL", "SUVR").

technology

Underlying method used (e.g., "ELISA", "PET", "Simoa").

vendor

Manufacturer of the assay.

platform

Commercial platform or product line name.

description

Descriptive label for the assay.

See also

Examples

assays()
#> # A tibble: 20 × 11
#>    assay_id     analyte_id modality matrix  lloq    uloq unit  technology vendor
#>    <chr>        <chr>      <chr>    <chr>  <dbl>   <dbl> <chr> <chr>      <chr> 
#>  1 innotest-ab… CHEBI:646… fluid    CSF    225    1452   pg/mL ELISA      Fujir…
#>  2 innotest-ab… CHEBI:646… fluid    CSF      3.4 10000   pg/mL ELISA      Fujir…
#>  3 lumipulse-g… CHEBI:646… fluid    CSF     NA      NA   pg/mL CLEIA      Fujir…
#>  4 lumipulse-g… CHEBI:646… fluid    CSF     NA      NA   pg/mL CLEIA      Fujir…
#>  5 euroimmun-a… CHEBI:646… fluid    CSF      6.5   641   pg/mL ELISA      EUROI…
#>  6 euroimmun-a… CHEBI:646… fluid    CSF    153   11000   pg/mL ELISA      EUROI…
#>  7 innogenetic… CHEBI:646… fluid    CSF     NA      NA   pg/mL ELISA      Innog…
#>  8 innogenetic… CHEBI:646… fluid    CSF     NA      NA   pg/mL ELISA      Innog…
#>  9 msd-ab42-csf CHEBI:646… fluid    CSF      3    1500   pg/mL ECLIA      MSD   
#> 10 msd-ab40-csf CHEBI:646… fluid    CSF     10    2500   pg/mL ECLIA      MSD   
#> 11 xmap-ab42-c… CHEBI:646… fluid    CSF     25    1700   pg/mL xMAP       Fujir…
#> 12 xmap-ab40-c… CHEBI:646… fluid    CSF     20    2500   pg/mL xMAP       Fujir…
#> 13 elecsys-pta… P10636-pT… fluid    CSF      0.5  1200   pg/mL ECLIA      Roche 
#> 14 lumipulse-p… P10636-pT… fluid    CSF      1    1500   pg/mL CLEIA      Fujir…
#> 15 simoa-ptau1… P10636-pT… fluid    plasma   0.2   500   pg/mL Simoa      Quant…
#> 16 simoa-nfl-p… P07196     fluid    plasma   2    1800   pg/mL Simoa      Quant…
#> 17 simoa-nfl-c… P07196     fluid    CSF      1    2500   pg/mL Simoa      Quant…
#> 18 simoa-gfap-… P14136     fluid    plasma   0.5  1200   pg/mL Simoa      Quant…
#> 19 pet-tau-flo… P10636     imaging  brain    0.3     3   SUVR  PET        AVID  
#> 20 pet-amyloid… CHEBI:646… imaging  brain    0.3     2.5 SUVR  PET        AVID  
#> # ℹ 2 more variables: platform <chr>, description <chr>