Retrieve biomarker cut-point values or metadata

cutpoints(): Provides access to a curated table of biomarker cut-points used to dichotomise or categorise continuous biomarker measurements in clinical or research settings.

cutpts() returns only the numeric cut-point(s) associated with that identifier. These thresholds are typically used for classifying biomarker values into diagnostic categories such as positive, negative, or indeterminate, see bin_by().

Usage

cutpoints()

cutpts(cutpt_id)

Arguments

cutpt_id

A character scalar specifying the identifier of the cut-point to retrieve.

Value

For cutpoints(): a tibble with columns cutpt_id, bmk_id, cutpts, unit, assay_id, intended_use, source, and notes.

For cutpts(): numeric value(s) corresponding to the threshold(s) defined for that biomarker-assay combination.

See also

Examples

# Return the full cut-point table with metadata
cutpoints()
#> # A tibble: 9 × 8
#>   cutpt_id             bmk_id cutpoints unit  assay_id intended_use source notes
#>   <chr>                <chr>      <dbl> <chr> <chr>    <chr>        <chr>  <chr>
#> 1 ab42-csf-elecsys-wi… ab42   1092      pg/mL elecsys… research     10.10… Deri…
#> 2 ptau181-ab42-ratio-… ptau1…    0.025  ratio elecsys… research     10.10… Deri…
#> 3 ttau-ab42-ratio-csf… ttau-…    0.27   ratio elecsys… research     10.10… Deri…
#> 4 ab42-csf-elecsys-ha… ab42   1100      pg/mL elecsys… research     10.10… NA   
#> 5 ptau181-ab42-ratio-… ptau1…    0.0215 ratio elecsys… research     10.10… NA   
#> 6 ttau-ab42-ratio-csf… ttau-…    0.241  ratio elecsys… research     10.10… NA   
#> 7 NA                   NA       NA      NA    NA       NA           NA     NA   
#> 8 ab42-innotest        ab42     NA      NA    innotes… diagnosis    PMID:… CSF …
#> 9 ptau181-elecsys      ptau1…   NA      NA    elecsys… diagnosis    PMID:… Elec…

# Retrieve the numeric cut-point(s) for a specific identifier
cutpts("ab42-csf-elecsys-willemse2018")
#> [1] 1092
cutpts("ptau181-ab42-ratio-csf-elecsys-willemse2018")
#> [1] 0.025