Limits of quantification per method

loq() is a convenience function that retrieves the limits of quantification from one of the assays in assays().

Usage

loq(assay_id)

Arguments

assay_id

A method identifier. One of the values in the assay_id variable in assays().

Value

A named numeric vector of two elements: lloq, lower, and uloq, upper, limits of quantification.

Examples

# Limits of quantification of Fujirebio INNOTEST β-Amyloid(1-42) CSF ELISA
# Check `assays()` for available methods and associated identifiers.
loq(assay_id = "innotest-ab42-csf")
#> lloq uloq 
#>  225 1452 

# Likewise, for Beta-Amyloid(1-40)
loq(assay_id = "innotest-ab40-csf")
#>    lloq    uloq 
#>     3.4 10000.0