Retrieve biomarker cut-point values or metadata

cutpoints(): Provides access to a curated table of biomarker cut-points used to dichotomise or categorise continuous biomarker measurements in clinical or research settings.

cutpts() returns only the numeric cut-point(s) associated with that identifier. These thresholds are typically used for classifying biomarker values into diagnostic categories such as positive, negative, or indeterminate, see cut_by().

Usage

cutpoints()

cutpts(x, direction = c("increasing", "decreasing"))

Arguments

x

A character scalar specifying the identifier of the cut-point to retrieve, or a numeric vector of cut-point values.

direction

Does positivity increase with x ("increasing") or does it decrease with x ("decreasing").

Value

For cutpoints(): a tibble with columns cutpt_id, bmk_id, cutpts, unit, assay_id, intended_use, source, and notes.

For cutpts(): numeric value(s) corresponding to the threshold(s) defined for that biomarker-assay combination.

See also

Examples

# Return the full cut-point table with metadata
cutpoints()
#> # A tibble: 8 × 9
#>   cutpt_id   bmk_id cutpoints unit  direction assay_id intended_use source notes
#>   <chr>      <chr>  <list>    <chr> <chr>     <chr>    <chr>        <chr>  <chr>
#> 1 ab42-csf-… ab42   <dbl [1]> pg/mL decreasi… elecsys… research     10.10… Deri…
#> 2 ptau181-a… ptau1… <dbl [1]> ratio increasi… elecsys… research     10.10… Deri…
#> 3 ttau-ab42… ttau_… <dbl [1]> ratio increasi… elecsys… research     10.10… Deri…
#> 4 ab42-csf-… ab42   <dbl [1]> pg/mL decreasi… elecsys… research     10.10… NA   
#> 5 ptau181-a… ptau1… <dbl [1]> ratio increasi… elecsys… research     10.10… NA   
#> 6 ttau-ab42… ttau_… <dbl [1]> ratio increasi… elecsys… research     10.10… NA   
#> 7 ab42-inno… ab42   <dbl [1]> NA    increasi… innotes… diagnosis    PMID:… CSF …
#> 8 ptau181-e… ptau1… <dbl [1]> NA    increasi… elecsys… diagnosis    PMID:… Elec…

# One cut-point
cutpts(5)
#> (-) < 5 <= (+) 

# Two cut-points
cutpts(c(2, 6))
#> (-) < 2 <= (~) < 6 <= (+) 

# Two cut-points but increasing values of `x` mean a negative result.
cutpts(c(2, 6), "decreasing")
#> (+) <= 2 < (~) <= 6 < (-) 

# Retrieve the numeric cut-point(s) for a specific identifier
cutpts("ab42-csf-elecsys-willemse2018")
#> (+) <= 1092 < (-) 
cutpts("ptau181-ab42-ratio-csf-elecsys-willemse2018")
#> (-) < 0.025 <= (+)